LPS-53617 EditArticleAction: move to Portlet class (no logic changes)
