--- portal-impl/src/com/liferay/portlet/journal/action/EditArticleAction.java
@@ -1,671 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portlet.journal.action;
-
-import com.liferay.portal.LocaleException;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.portlet.LiferayWindowState;
-import com.liferay.portal.kernel.servlet.SessionErrors;
-import com.liferay.portal.kernel.servlet.SessionMessages;
-import com.liferay.portal.kernel.upload.LiferayFileItemException;
-import com.liferay.portal.kernel.upload.UploadException;
-import com.liferay.portal.kernel.upload.UploadPortletRequest;
-import com.liferay.portal.kernel.util.Constants;
-import com.liferay.portal.kernel.util.HtmlUtil;
-import com.liferay.portal.kernel.util.HttpUtil;
-import com.liferay.portal.kernel.util.LocalizationUtil;
-import com.liferay.portal.kernel.util.MapUtil;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.workflow.WorkflowConstants;
-import com.liferay.portal.model.Layout;
-import com.liferay.portal.model.TrashedModel;
-import com.liferay.portal.security.auth.PrincipalException;
-import com.liferay.portal.service.ServiceContext;
-import com.liferay.portal.service.ServiceContextFactory;
-import com.liferay.portal.struts.PortletAction;
-import com.liferay.portal.theme.ThemeDisplay;
-import com.liferay.portal.util.PortalUtil;
-import com.liferay.portal.util.WebKeys;
-import com.liferay.portlet.PortletURLFactoryUtil;
-import com.liferay.portlet.PortletURLImpl;
-import com.liferay.portlet.asset.AssetCategoryException;
-import com.liferay.portlet.asset.AssetTagException;
-import com.liferay.portlet.documentlibrary.DuplicateFileException;
-import com.liferay.portlet.documentlibrary.FileSizeException;
-import com.liferay.portlet.dynamicdatamapping.NoSuchStructureException;
-import com.liferay.portlet.dynamicdatamapping.NoSuchTemplateException;
-import com.liferay.portlet.dynamicdatamapping.StorageFieldRequiredException;
-import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
-import com.liferay.portlet.dynamicdatamapping.service.DDMStructureLocalServiceUtil;
-import com.liferay.portlet.journal.ArticleContentException;
-import com.liferay.portlet.journal.ArticleContentSizeException;
-import com.liferay.portlet.journal.ArticleDisplayDateException;
-import com.liferay.portlet.journal.ArticleExpirationDateException;
-import com.liferay.portlet.journal.ArticleIdException;
-import com.liferay.portlet.journal.ArticleSmallImageNameException;
-import com.liferay.portlet.journal.ArticleSmallImageSizeException;
-import com.liferay.portlet.journal.ArticleTitleException;
-import com.liferay.portlet.journal.ArticleVersionException;
-import com.liferay.portlet.journal.DuplicateArticleIdException;
-import com.liferay.portlet.journal.NoSuchArticleException;
-import com.liferay.portlet.journal.asset.JournalArticleAssetRenderer;
-import com.liferay.portlet.journal.model.JournalArticle;
-import com.liferay.portlet.journal.service.JournalArticleServiceUtil;
-import com.liferay.portlet.journal.service.JournalContentSearchLocalServiceUtil;
-import com.liferay.portlet.journal.util.JournalUtil;
-import com.liferay.portlet.trash.util.TrashUtil;
-
-import java.io.File;
-
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-
-import javax.portlet.ActionRequest;
-import javax.portlet.ActionResponse;
-import javax.portlet.PortletConfig;
-import javax.portlet.PortletPreferences;
-import javax.portlet.PortletRequest;
-import javax.portlet.PortletURL;
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
-import javax.portlet.WindowState;
-
-import org.apache.struts.action.ActionForm;
-import org.apache.struts.action.ActionForward;
-import org.apache.struts.action.ActionMapping;
-
-/**
- * @author Brian Wing Shun Chan
- * @author Raymond Augé
- * @author Eduardo Lundgren
- * @author Juan Fernández
- * @author Levente Hudák
- */
-public class EditArticleAction extends PortletAction {
-
-	public static final String VERSION_SEPARATOR = "_version_";
-
-	@Override
-	public void processAction(
-			ActionMapping actionMapping, ActionForm actionForm,
-			PortletConfig portletConfig, ActionRequest actionRequest,
-			ActionResponse actionResponse)
-		throws Exception {
-
-		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
-
-		JournalArticle article = null;
-		String oldUrlTitle = StringPool.BLANK;
-
-		try {
-			UploadException uploadException =
-				(UploadException)actionRequest.getAttribute(
-					WebKeys.UPLOAD_EXCEPTION);
-
-			if (uploadException != null) {
-				if (uploadException.isExceededLiferayFileItemSizeLimit()) {
-					throw new LiferayFileItemException();
-				}
-				else if (uploadException.isExceededSizeLimit()) {
-					throw new ArticleContentSizeException();
-				}
-
-				throw new PortalException(uploadException.getCause());
-			}
-			else if (Validator.isNull(cmd)) {
-				return;
-			}
-			else if (cmd.equals(Constants.ADD) ||
-					 cmd.equals(Constants.PREVIEW) ||
-					 cmd.equals(Constants.UPDATE)) {
-
-				Object[] contentAndImages = updateArticle(actionRequest);
-
-				article = (JournalArticle)contentAndImages[0];
-				oldUrlTitle = ((String)contentAndImages[1]);
-			}
-			else if (cmd.equals(Constants.DELETE)) {
-				deleteArticles(actionRequest, false);
-			}
-			else if (cmd.equals(Constants.EXPIRE)) {
-				expireArticles(actionRequest);
-			}
-			else if (cmd.equals(Constants.MOVE_TO_TRASH)) {
-				deleteArticles(actionRequest, true);
-			}
-			else if (cmd.equals(Constants.SUBSCRIBE)) {
-				subscribeStructure(actionRequest);
-			}
-			else if (cmd.equals(Constants.UNSUBSCRIBE)) {
-				unsubscribeStructure(actionRequest);
-			}
-
-			String redirect = ParamUtil.getString(actionRequest, "redirect");
-
-			int workflowAction = ParamUtil.getInteger(
-				actionRequest, "workflowAction",
-				WorkflowConstants.ACTION_PUBLISH);
-
-			String portletId = HttpUtil.getParameter(redirect, "p_p_id", false);
-
-			String namespace = PortalUtil.getPortletNamespace(portletId);
-
-			if (Validator.isNotNull(oldUrlTitle)) {
-				String oldRedirectParam = namespace + "redirect";
-
-				String oldRedirect = HttpUtil.getParameter(
-					redirect, oldRedirectParam, false);
-
-				if (Validator.isNotNull(oldRedirect)) {
-					String newRedirect = HttpUtil.decodeURL(oldRedirect);
-
-					newRedirect = StringUtil.replace(
-						newRedirect, oldUrlTitle, article.getUrlTitle());
-					newRedirect = StringUtil.replace(
-						newRedirect, oldRedirectParam, "redirect");
-
-					redirect = StringUtil.replace(
-						redirect, oldRedirect, newRedirect);
-				}
-			}
-
-			if (cmd.equals(Constants.DELETE) &&
-				!ActionUtil.hasArticle(actionRequest)) {
-
-				ThemeDisplay themeDisplay =
-					(ThemeDisplay)actionRequest.getAttribute(
-						WebKeys.THEME_DISPLAY);
-
-				PortletURL portletURL = PortletURLFactoryUtil.create(
-					actionRequest, portletConfig.getPortletName(),
-					themeDisplay.getPlid(), PortletRequest.RENDER_PHASE);
-
-				redirect = portletURL.toString();
-			}
-
-			if ((article != null) &&
-				(workflowAction == WorkflowConstants.ACTION_SAVE_DRAFT)) {
-
-				redirect = getSaveAndContinueRedirect(
-					portletConfig, actionRequest, article, redirect);
-
-				if (cmd.equals(Constants.PREVIEW)) {
-					SessionMessages.add(actionRequest, "previewRequested");
-
-					hideDefaultSuccessMessage(actionRequest);
-				}
-
-				sendRedirect(actionRequest, actionResponse, redirect);
-			}
-			else {
-				WindowState windowState = actionRequest.getWindowState();
-
-				if (!windowState.equals(LiferayWindowState.POP_UP)) {
-					sendRedirect(actionRequest, actionResponse, redirect);
-				}
-				else {
-					redirect = PortalUtil.escapeRedirect(redirect);
-
-					if (Validator.isNotNull(redirect)) {
-						if (cmd.equals(Constants.ADD) && (article != null)) {
-							redirect = HttpUtil.addParameter(
-								redirect, namespace + "className",
-								JournalArticle.class.getName());
-							redirect = HttpUtil.addParameter(
-								redirect, namespace + "classPK",
-								JournalArticleAssetRenderer.getClassPK(
-									article));
-						}
-
-						actionResponse.sendRedirect(redirect);
-					}
-				}
-			}
-		}
-		catch (Exception e) {
-			if (e instanceof NoSuchArticleException ||
-				e instanceof NoSuchStructureException ||
-				e instanceof NoSuchTemplateException ||
-				e instanceof PrincipalException) {
-
-				SessionErrors.add(actionRequest, e.getClass());
-
-				setForward(actionRequest, "portlet.journal.error");
-			}
-			else if (e instanceof ArticleContentException ||
-					 e instanceof ArticleContentSizeException ||
-					 e instanceof ArticleDisplayDateException ||
-					 e instanceof ArticleExpirationDateException ||
-					 e instanceof ArticleIdException ||
-					 e instanceof ArticleSmallImageNameException ||
-					 e instanceof ArticleSmallImageSizeException ||
-					 e instanceof ArticleTitleException ||
-					 e instanceof ArticleVersionException ||
-					 e instanceof DuplicateArticleIdException ||
-					 e instanceof DuplicateFileException ||
-					 e instanceof FileSizeException ||
-					 e instanceof LiferayFileItemException ||
-					 e instanceof StorageFieldRequiredException) {
-
-				SessionErrors.add(actionRequest, e.getClass());
-			}
-			else if (e instanceof AssetCategoryException ||
-					 e instanceof AssetTagException ||
-					 e instanceof LocaleException) {
-
-				SessionErrors.add(actionRequest, e.getClass(), e);
-			}
-			else {
-				throw e;
-			}
-		}
-	}
-
-	@Override
-	public ActionForward render(
-			ActionMapping actionMapping, ActionForm actionForm,
-			PortletConfig portletConfig, RenderRequest renderRequest,
-			RenderResponse renderResponse)
-		throws Exception {
-
-		try {
-			ActionUtil.getArticle(renderRequest);
-		}
-		catch (NoSuchArticleException nsae) {
-
-			// Let this slide because the user can manually input a article id
-			// for a new article that does not yet exist.
-
-		}
-		catch (Exception e) {
-			if (//e instanceof NoSuchArticleException ||
-				e instanceof PrincipalException) {
-
-				SessionErrors.add(renderRequest, e.getClass());
-
-				return actionMapping.findForward("portlet.journal.error");
-			}
-			else {
-				throw e;
-			}
-		}
-
-		return actionMapping.findForward(
-			getForward(renderRequest, "portlet.journal.edit_article"));
-	}
-
-	protected void deleteArticles(
-			ActionRequest actionRequest, boolean moveToTrash)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		String[] deleteArticleIds = null;
-
-		String articleId = ParamUtil.getString(actionRequest, "articleId");
-
-		if (Validator.isNotNull(articleId)) {
-			deleteArticleIds = new String[] {articleId};
-		}
-		else {
-			deleteArticleIds = StringUtil.split(
-				ParamUtil.getString(actionRequest, "articleIds"));
-		}
-
-		List<TrashedModel> trashedModels = new ArrayList<>();
-
-		for (String deleteArticleId : deleteArticleIds) {
-			if (moveToTrash) {
-				JournalArticle article =
-					JournalArticleServiceUtil.moveArticleToTrash(
-						themeDisplay.getScopeGroupId(),
-						HtmlUtil.unescape(deleteArticleId));
-
-				trashedModels.add(article);
-			}
-			else {
-				ActionUtil.deleteArticle(
-					actionRequest, HtmlUtil.unescape(deleteArticleId));
-			}
-		}
-
-		if (moveToTrash && !trashedModels.isEmpty()) {
-			TrashUtil.addTrashSessionMessages(actionRequest, trashedModels);
-
-			hideDefaultSuccessMessage(actionRequest);
-		}
-	}
-
-	protected void expireArticles(ActionRequest actionRequest)
-		throws Exception {
-
-		String articleId = ParamUtil.getString(actionRequest, "articleId");
-
-		if (Validator.isNotNull(articleId)) {
-			ActionUtil.expireArticle(actionRequest, articleId);
-		}
-		else {
-			String[] expireArticleIds = StringUtil.split(
-				ParamUtil.getString(actionRequest, "expireArticleIds"));
-
-			for (String expireArticleId : expireArticleIds) {
-				ActionUtil.expireArticle(
-					actionRequest, HtmlUtil.unescape(expireArticleId));
-			}
-		}
-	}
-
-	protected String getSaveAndContinueRedirect(
-			PortletConfig portletConfig, ActionRequest actionRequest,
-			JournalArticle article, String redirect)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		String referringPortletResource = ParamUtil.getString(
-			actionRequest, "referringPortletResource");
-
-		PortletURLImpl portletURL = new PortletURLImpl(
-			actionRequest, portletConfig.getPortletName(),
-			themeDisplay.getPlid(), PortletRequest.RENDER_PHASE);
-
-		portletURL.setParameter("struts_action", "/journal/edit_article");
-		portletURL.setParameter("redirect", redirect, false);
-		portletURL.setParameter(
-			"referringPortletResource", referringPortletResource, false);
-		portletURL.setParameter(
-			"resourcePrimKey", String.valueOf(article.getResourcePrimKey()),
-			false);
-		portletURL.setParameter(
-			"groupId", String.valueOf(article.getGroupId()), false);
-		portletURL.setParameter(
-			"folderId", String.valueOf(article.getFolderId()), false);
-		portletURL.setParameter("articleId", article.getArticleId(), false);
-		portletURL.setParameter(
-			"version", String.valueOf(article.getVersion()), false);
-		portletURL.setWindowState(actionRequest.getWindowState());
-
-		return portletURL.toString();
-	}
-
-	protected void subscribeStructure(ActionRequest actionRequest)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		long ddmStructureId = ParamUtil.getLong(
-			actionRequest, "ddmStructureId");
-
-		JournalArticleServiceUtil.subscribeStructure(
-			themeDisplay.getScopeGroupId(), themeDisplay.getUserId(),
-			ddmStructureId);
-	}
-
-	protected void unsubscribeStructure(ActionRequest actionRequest)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		long ddmStructureId = ParamUtil.getLong(
-			actionRequest, "ddmStructureId");
-
-		JournalArticleServiceUtil.unsubscribeStructure(
-			themeDisplay.getScopeGroupId(), themeDisplay.getUserId(),
-			ddmStructureId);
-	}
-
-	protected Object[] updateArticle(ActionRequest actionRequest)
-		throws Exception {
-
-		UploadPortletRequest uploadPortletRequest =
-			PortalUtil.getUploadPortletRequest(actionRequest);
-
-		if (_log.isDebugEnabled()) {
-			_log.debug(
-				"Updating article " +
-					MapUtil.toString(uploadPortletRequest.getParameterMap()));
-		}
-
-		String cmd = ParamUtil.getString(uploadPortletRequest, Constants.CMD);
-
-		long groupId = ParamUtil.getLong(uploadPortletRequest, "groupId");
-		long folderId = ParamUtil.getLong(uploadPortletRequest, "folderId");
-		long classNameId = ParamUtil.getLong(
-			uploadPortletRequest, "classNameId");
-		long classPK = ParamUtil.getLong(uploadPortletRequest, "classPK");
-
-		String articleId = ParamUtil.getString(
-			uploadPortletRequest, "articleId");
-
-		boolean autoArticleId = ParamUtil.getBoolean(
-			uploadPortletRequest, "autoArticleId");
-		double version = ParamUtil.getDouble(uploadPortletRequest, "version");
-
-		Map<Locale, String> titleMap = LocalizationUtil.getLocalizationMap(
-			actionRequest, "title");
-		Map<Locale, String> descriptionMap =
-			LocalizationUtil.getLocalizationMap(actionRequest, "description");
-
-		ServiceContext serviceContext = ServiceContextFactory.getInstance(
-			JournalArticle.class.getName(), uploadPortletRequest);
-
-		String ddmStructureKey = ParamUtil.getString(
-			uploadPortletRequest, "ddmStructureKey");
-
-		DDMStructure ddmStructure = DDMStructureLocalServiceUtil.getStructure(
-			PortalUtil.getSiteGroupId(groupId),
-			PortalUtil.getClassNameId(JournalArticle.class), ddmStructureKey,
-			true);
-
-		Object[] contentAndImages = ActionUtil.getContentAndImages(
-			ddmStructure, serviceContext);
-
-		String content = (String)contentAndImages[0];
-		Map<String, byte[]> images =
-			(HashMap<String, byte[]>)contentAndImages[1];
-
-		Boolean fileItemThresholdSizeExceeded =
-			(Boolean)uploadPortletRequest.getAttribute(
-				WebKeys.FILE_ITEM_THRESHOLD_SIZE_EXCEEDED);
-
-		if ((fileItemThresholdSizeExceeded != null) &&
-			fileItemThresholdSizeExceeded.booleanValue()) {
-
-			throw new ArticleContentSizeException();
-		}
-
-		String ddmTemplateKey = ParamUtil.getString(
-			uploadPortletRequest, "ddmTemplateKey");
-		String layoutUuid = ParamUtil.getString(
-			uploadPortletRequest, "layoutUuid");
-
-		Layout targetLayout = JournalUtil.getArticleLayout(layoutUuid, groupId);
-
-		if (targetLayout == null) {
-			layoutUuid = null;
-		}
-
-		int displayDateMonth = ParamUtil.getInteger(
-			uploadPortletRequest, "displayDateMonth");
-		int displayDateDay = ParamUtil.getInteger(
-			uploadPortletRequest, "displayDateDay");
-		int displayDateYear = ParamUtil.getInteger(
-			uploadPortletRequest, "displayDateYear");
-		int displayDateHour = ParamUtil.getInteger(
-			uploadPortletRequest, "displayDateHour");
-		int displayDateMinute = ParamUtil.getInteger(
-			uploadPortletRequest, "displayDateMinute");
-		int displayDateAmPm = ParamUtil.getInteger(
-			uploadPortletRequest, "displayDateAmPm");
-
-		if (displayDateAmPm == Calendar.PM) {
-			displayDateHour += 12;
-		}
-
-		int expirationDateMonth = ParamUtil.getInteger(
-			uploadPortletRequest, "expirationDateMonth");
-		int expirationDateDay = ParamUtil.getInteger(
-			uploadPortletRequest, "expirationDateDay");
-		int expirationDateYear = ParamUtil.getInteger(
-			uploadPortletRequest, "expirationDateYear");
-		int expirationDateHour = ParamUtil.getInteger(
-			uploadPortletRequest, "expirationDateHour");
-		int expirationDateMinute = ParamUtil.getInteger(
-			uploadPortletRequest, "expirationDateMinute");
-		int expirationDateAmPm = ParamUtil.getInteger(
-			uploadPortletRequest, "expirationDateAmPm");
-		boolean neverExpire = ParamUtil.getBoolean(
-			uploadPortletRequest, "neverExpire");
-
-		if (expirationDateAmPm == Calendar.PM) {
-			expirationDateHour += 12;
-		}
-
-		int reviewDateMonth = ParamUtil.getInteger(
-			uploadPortletRequest, "reviewDateMonth");
-		int reviewDateDay = ParamUtil.getInteger(
-			uploadPortletRequest, "reviewDateDay");
-		int reviewDateYear = ParamUtil.getInteger(
-			uploadPortletRequest, "reviewDateYear");
-		int reviewDateHour = ParamUtil.getInteger(
-			uploadPortletRequest, "reviewDateHour");
-		int reviewDateMinute = ParamUtil.getInteger(
-			uploadPortletRequest, "reviewDateMinute");
-		int reviewDateAmPm = ParamUtil.getInteger(
-			uploadPortletRequest, "reviewDateAmPm");
-		boolean neverReview = ParamUtil.getBoolean(
-			uploadPortletRequest, "neverReview");
-
-		if (reviewDateAmPm == Calendar.PM) {
-			reviewDateHour += 12;
-		}
-
-		boolean indexable = ParamUtil.getBoolean(
-			uploadPortletRequest, "indexable");
-
-		boolean smallImage = ParamUtil.getBoolean(
-			uploadPortletRequest, "smallImage");
-		String smallImageURL = ParamUtil.getString(
-			uploadPortletRequest, "smallImageURL");
-		File smallFile = uploadPortletRequest.getFile("smallFile");
-
-		String articleURL = ParamUtil.getString(
-			uploadPortletRequest, "articleURL");
-
-		JournalArticle article = null;
-		String oldUrlTitle = StringPool.BLANK;
-
-		if (cmd.equals(Constants.ADD)) {
-
-			// Add article
-
-			article = JournalArticleServiceUtil.addArticle(
-				groupId, folderId, classNameId, classPK, articleId,
-				autoArticleId, titleMap, descriptionMap, content,
-				ddmStructureKey, ddmTemplateKey, layoutUuid, displayDateMonth,
-				displayDateDay, displayDateYear, displayDateHour,
-				displayDateMinute, expirationDateMonth, expirationDateDay,
-				expirationDateYear, expirationDateHour, expirationDateMinute,
-				neverExpire, reviewDateMonth, reviewDateDay, reviewDateYear,
-				reviewDateHour, reviewDateMinute, neverReview, indexable,
-				smallImage, smallImageURL, smallFile, images, articleURL,
-				serviceContext);
-		}
-		else {
-
-			// Update article
-
-			article = JournalArticleServiceUtil.getArticle(
-				groupId, articleId, version);
-
-			String tempOldUrlTitle = article.getUrlTitle();
-
-			if (cmd.equals(Constants.PREVIEW) || cmd.equals(Constants.UPDATE)) {
-				article = JournalArticleServiceUtil.updateArticle(
-					groupId, folderId, articleId, version, titleMap,
-					descriptionMap, content, ddmStructureKey, ddmTemplateKey,
-					layoutUuid, displayDateMonth, displayDateDay,
-					displayDateYear, displayDateHour, displayDateMinute,
-					expirationDateMonth, expirationDateDay, expirationDateYear,
-					expirationDateHour, expirationDateMinute, neverExpire,
-					reviewDateMonth, reviewDateDay, reviewDateYear,
-					reviewDateHour, reviewDateMinute, neverReview, indexable,
-					smallImage, smallImageURL, smallFile, images, articleURL,
-					serviceContext);
-			}
-
-			if (!tempOldUrlTitle.equals(article.getUrlTitle())) {
-				oldUrlTitle = tempOldUrlTitle;
-			}
-		}
-
-		// Recent articles
-
-		JournalUtil.addRecentArticle(actionRequest, article);
-
-		// Journal content
-
-		PortletPreferences portletPreferences = getStrictPortletSetup(
-			actionRequest);
-
-		if (portletPreferences != null) {
-			portletPreferences.setValue(
-				"groupId", String.valueOf(article.getGroupId()));
-			portletPreferences.setValue("articleId", article.getArticleId());
-
-			portletPreferences.store();
-
-			String portletResource = ParamUtil.getString(
-				actionRequest, "portletResource");
-
-			updateContentSearch(
-				actionRequest, portletResource, article.getArticleId());
-		}
-
-		return new Object[] {article, oldUrlTitle};
-	}
-
-	protected void updateContentSearch(
-			ActionRequest actionRequest, String portletResource,
-			String articleId)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		Layout layout = themeDisplay.getLayout();
-
-		JournalContentSearchLocalServiceUtil.updateContentSearch(
-			layout.getGroupId(), layout.isPrivateLayout(), layout.getLayoutId(),
-			portletResource, articleId, true);
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		EditArticleAction.class);
-
-}
